Move all TileLayer code to the tile_layer/ directory. Add Coords to the exports since it's needed for custom TileProvider implementations
