+++ lib/flutter_map.dart
@@ -32,12 +32,13 @@ export 'package:flutter_map/src/layer/marker_layer.dart';
 export 'package:flutter_map/src/layer/overlay_image_layer.dart';
 export 'package:flutter_map/src/layer/polygon_layer.dart';
 export 'package:flutter_map/src/layer/polyline_layer.dart';
+export 'package:flutter_map/src/layer/tile_layer/coords.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile.dart';
-export 'package:flutter_map/src/layer/tile_builder/tile_builder.dart';
+export 'package:flutter_map/src/layer/tile_layer/tile_builder.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_layer.dart';
-export 'package:flutter_map/src/layer/tile_provider/file_tile_provider_io.dart'
-    if (dart.library.html) 'package:flutter_map/src/layer/tile_provider/file_tile_provider_web.dart';
-export 'package:flutter_map/src/layer/tile_provider/tile_provider.dart';
+export 'package:flutter_map/src/layer/tile_layer/tile_provider/file_tile_provider_io.dart'
+    if (dart.library.html) 'package:flutter_map/src/layer/tile_layer/tile_provider/file_tile_provider_web.dart';
+export 'package:flutter_map/src/layer/tile_layer/tile_provider/tile_provider.dart';
 export 'package:flutter_map/src/plugins/plugin.dart';
 
 /// Renders a map composed of a list of layers powered by [LayerOptions].
@@ -146,8 +147,11 @@ abstract class MapController {
   double get rotation;
 
   Stream<MapEvent> get mapEventStream;
+
   StreamSink<MapEvent> get mapEventSink;
+
   set state(MapState state);
+
   void dispose();
 
   LatLng? pointToLatLng(CustomPoint point);
