+++ lib/src/layer/tile_layer/tile.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:flutter_map/src/layer/tile_layer/coords.dart';
 
 typedef TileReady = void Function(
     Coords<double> coords, dynamic error, Tile tile);
