+++ lib/src/layer/tile_layer/tile_manager.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/core/bounds.dart';
-import 'package:flutter_map/src/layer/tile_layer/coords.dart';
 import 'package:tuple/tuple.dart';
 
 class TileManager {
